backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
requests==2.32.5
openai==1.101.0
langchain==0.3.27
langchain-community==0.3.27
pydantic==2.11.7
python-dotenv==1.1.1
ollama==0.5.3
